unpack.py: add SRC_URI parameter unpack=<bool> (default: true)
* This allows to download compressed files without extracting them
* Use case: gcj requires ecj.jar, which must be downloaded separately
  and put into the gcc source directory before configure gets executed.

CC: Chris Larson <chris_larson@mentor.com>

+++ lib/oe/unpack.py
@@ -47,8 +47,20 @@ def subprocess_setup():
     # non-Python subprocesses expect.
     signal.signal(signal.SIGPIPE, signal.SIG_DFL)
 
-def unpack_file(file, destdir, dos=False, env=None):
+def unpack_file(file, destdir, parameters, env=None):
     import subprocess, shutil
+
+    try:
+        dos = to_boolean(parameters.get("dos"), False)
+    except ValueError, exc:
+        bb.fatal("Invalid value for 'dos' parameter for %s: %s" %
+                 (filename, parameters.get("dos")))
+
+    try:
+        unpack = to_boolean(parameters.get("unpack"), True)
+    except ValueError, exc:
+        bb.fatal("Invalid value for 'unpack' parameter for %s: %s" %
+                 (filename, parameters.get("unpack")))
 
     dest = os.path.join(destdir, os.path.basename(file))
     if os.path.exists(dest):
@@ -56,38 +68,41 @@ def unpack_file(file, destdir, dos=False, env=None):
             return True
 
     cmd = None
-    if file.endswith('.tar'):
-        cmd = 'tar x --no-same-owner -f %s' % file
-    elif file.endswith('.tgz') or file.endswith('.tar.gz') or file.endswith('.tar.Z'):
-        cmd = 'tar xz --no-same-owner -f %s' % file
-    elif file.endswith('.tbz') or file.endswith('.tbz2') or file.endswith('.tar.bz2'):
-        cmd = 'bzip2 -dc %s | tar x --no-same-owner -f -' % file
-    elif file.endswith('.gz') or file.endswith('.Z') or file.endswith('.z'):
-        root, ext = os.path.splitext(file)
-        cmd = 'gzip -dc %s > %s' % (file, os.path.basename(root))
-    elif file.endswith('.bz2'):
-        root, ext = os.path.splitext(file)
-        cmd = 'bzip2 -dc %s > %s' % (file, os.path.basename(root))
-    elif file.endswith('.tar.xz'):
-        cmd = 'xz -dc %s | tar x --no-same-owner -f -' % file
-    elif file.endswith('.xz'):
-        root, ext = os.path.splitext(file)
-        cmd = 'xz -dc %s > %s' % (file, os.path.basename(root))
-    elif file.endswith('.tar.lz'):
-        cmd = 'lzip -dc %s | tar x --no-same-owner -f -' % file
-    elif file.endswith('.lz'):
-        root, ext = os.path.splitext(file)
-        cmd = 'lzip -dc %s > %s' % (file, os.path.basename(root))
-    elif file.endswith('.zip') or file.endswith('.jar'):
-        cmd = 'unzip -q -o'
-        if dos:
-            cmd = '%s -a' % cmd
-        cmd = "%s '%s'" % (cmd, file)
-    elif os.path.isdir(file):
-        shutil.rmtree(dest, True)
-        shutil.copytree(file, dest, True)
-    else:
-        shutil.copy2(file, dest)
+    if unpack:
+        if file.endswith('.tar'):
+            cmd = 'tar x --no-same-owner -f %s' % file
+        elif file.endswith('.tgz') or file.endswith('.tar.gz') or file.endswith('.tar.Z'):
+            cmd = 'tar xz --no-same-owner -f %s' % file
+        elif file.endswith('.tbz') or file.endswith('.tbz2') or file.endswith('.tar.bz2'):
+            cmd = 'bzip2 -dc %s | tar x --no-same-owner -f -' % file
+        elif file.endswith('.gz') or file.endswith('.Z') or file.endswith('.z'):
+            root, ext = os.path.splitext(file)
+            cmd = 'gzip -dc %s > %s' % (file, os.path.basename(root))
+        elif file.endswith('.bz2'):
+            root, ext = os.path.splitext(file)
+            cmd = 'bzip2 -dc %s > %s' % (file, os.path.basename(root))
+        elif file.endswith('.tar.xz'):
+            cmd = 'xz -dc %s | tar x --no-same-owner -f -' % file
+        elif file.endswith('.xz'):
+            root, ext = os.path.splitext(file)
+            cmd = 'xz -dc %s > %s' % (file, os.path.basename(root))
+        elif file.endswith('.tar.lz'):
+            cmd = 'lzip -dc %s | tar x --no-same-owner -f -' % file
+        elif file.endswith('.lz'):
+            root, ext = os.path.splitext(file)
+            cmd = 'lzip -dc %s > %s' % (file, os.path.basename(root))
+        elif file.endswith('.zip') or file.endswith('.jar'):
+            cmd = 'unzip -q -o'
+            if dos:
+                cmd = '%s -a' % cmd
+            cmd = "%s '%s'" % (cmd, file)
+
+    if not unpack or not cmd:
+        if os.path.isdir(file):
+            shutil.rmtree(dest, True)
+            shutil.copytree(file, dest, True)
+        else:
+            shutil.copy2(file, dest)
 
     if not cmd:
         return
